app/src/main/java/com/example/testapp/ui/main/MainViewModel.kt
package com.example.testapp.ui.main

import android.content.Context
import android.net.ConnectivityManager
import android.net.NetworkCapabilities
import android.util.Log
import androidx.lifecycle.ViewModel
import com.example.testapp.model.CakeModel
import com.example.testapp.repository.CakesRepository
import dagger.hilt.android.lifecycle.HiltViewModel
import io.reactivex.rxjava3.core.Observable
import javax.inject.Inject

@HiltViewModel
class MainViewModel @Inject constructor(
    private val cakesRepository: CakesRepository
) : ViewModel() {
    fun fetchCakeList():Observable<ArrayList<CakeModel>> {
        return cakesRepository.getCakes()
    }

    fun removeDuplicatesAndSort(cakeList: ArrayList<CakeModel>):List<CakeModel> {
        val uniqueCakes = cakeList.distinctBy { it.title }
        return uniqueCakes.sortedWith(compareBy { it.title })
    }

    fun isOnline(context: Context): Boolean {
        val connectivityManager =
            context.getSystemService(Context.CONNECTIVITY_SERVICE) as ConnectivityManager
        if (connectivityManager != null) {
            val capabilities =
                connectivityManager.getNetworkCapabilities(connectivityManager.activeNetwork)
            if (capabilities != null) {
                if (capabilities.hasTransport(NetworkCapabilities.TRANSPORT_CELLULAR)) {
                    Log.i("Internet", "NetworkCapabilities.TRANSPORT_CELLULAR")
                    return true
                } else if (capabilities.hasTransport(NetworkCapabilities.TRANSPORT_WIFI)) {
                    Log.i("Internet", "NetworkCapabilities.TRANSPORT_WIFI")
                    return true
                } else if (capabilities.hasTransport(NetworkCapabilities.TRANSPORT_ETHERNET)) {
                    Log.i("Internet", "NetworkCapabilities.TRANSPORT_ETHERNET")
                    return true
                }
            }
        }
        return false
    }
}